+++ src/main/java/electrolyte/greate/foundation/data/recipe/machine/GreateSpoutRecipes.java
@@ -58,7 +58,7 @@ public class GreateSpoutRecipes {
                     .recipeBuilder(Greate.id(String.format("cover_%s_%s_rubber", material.getName(), cableAmount)))
                     .EUt(VA[euT]).duration(100)
                     .inputItems(wirePrefix, material)
-                    .inputFluids(Rubber.getFluid(L * (long) insulationAmount))
+                    .inputFluids(Rubber.getFluid(L * insulationAmount))
                     .outputItems(cablePrefix, material)
                     .save(provider);
 
@@ -72,7 +72,7 @@ public class GreateSpoutRecipes {
                     .recipeBuilder(Greate.id(String.format("cover_%s_%s_silicone", material.getName(), cableAmount)))
                     .EUt(VA[euT]).duration(100)
                     .inputItems(wirePrefix, material)
-                    .inputFluids(SiliconeRubber.getFluid(L * (long) insulationAmount / 2))
+                    .inputFluids(SiliconeRubber.getFluid(L * insulationAmount / 2))
                     .outputItems(cablePrefix, material)
                     .save(provider);
 
@@ -86,7 +86,7 @@ public class GreateSpoutRecipes {
                     .recipeBuilder(Greate.id(String.format("cover_%s_%s_styrene_butadiene", material.getName(), cableAmount)))
                     .EUt(VA[euT]).duration(100)
                     .inputItems(wirePrefix, material)
-                    .inputFluids(StyreneButadieneRubber.getFluid(L * (long) insulationAmount / 4))
+                    .inputFluids(StyreneButadieneRubber.getFluid(L * insulationAmount / 4))
                     .outputItems(cablePrefix, material)
                     .save(provider);
         }
